refactor: extract iacStack to define resrouces separately
