+++ src/stack/deploy.ts
@@ -1,7 +1,7 @@
 import * as ros from '@alicloud/ros-cdk-core';
 import { ActionContext, ServerlessIac } from '../types';
 import { logger, rosStackDeploy } from '../common';
-import { IacStack } from './iacStack';
+import { RosStack } from './rosStack/rosStack';
 
 export const generateStackTemplate = (
   stackName: string,
@@ -9,7 +9,7 @@ export const generateStackTemplate = (
   context: ActionContext,
 ) => {
   const app = new ros.App();
-  new IacStack(app, iac, context);
+  new RosStack(app, iac, context);
 
   const assembly = app.synth();
   const stackArtifact = assembly.getStackByName(stackName);
